% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_master.R
\name{Blank_statistic}
\alias{Blank_statistic}
\title{Calculation of the basic library information based on standard sample-blank matrix}
\usage{
Blank_statistic(blank_mt, start = NA, end = NA)
}
\arguments{
\item{blank_mt}{A standard sample-blank matrix.}

\item{start}{The row index of the start sample in \strong{blank_mt}.}

\item{end}{The row index of the ending sample in \strong{blank_mt}.}
}
\value{
A vector contains:
\enumerate{
\item Number of samples (\strong{nSample});
\item Number of blanks (\strong{nBlank});
\item Sum number of samples and blanks (\strong{nTotal});
\item Ratio of \strong{nSample} to \strong{nBlank}.
}
}
\description{
Function \strong{Blank_statistic} records the start and ending sample of a library in a standard sample-blank matrix,
and calculates the number of samples and blanks and the sample-blank ratio (SBR, sample/blank).
}
\examples{
set.seed(2026)
blank_mt_test <- data.frame('Sample'=paste0('S', 1:32),
'B1'=sample(0:1, 32, replace = T, prob=c(0.75,0.25)),
'B2'=sample(0:1, 32, replace = T, prob=c(0.75,0.25)),
'B3'=sample(0:1, 32, replace = T, prob=c(0.75,0.25)),
'B4'=sample(0:1, 32, replace = T, prob=c(0.75,0.25)))
blank_sta_res <- Blank_statistic(blank_mt_test, start=1, end=16)
print(blank_sta_res)
}
