% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_master.R
\name{Lib_design}
\alias{Lib_design}
\title{Generate library design schemes}
\usage{
Lib_design(
  sample_list,
  div_step = 8,
  div_max_lib = 2,
  div_min_lib = 1,
  div_num_lib = NULL,
  div_print_summary = FALSE
)
}
\arguments{
\item{sample_list}{The original sample and blank list as data.frame, with each sample as one row.
It should include:
1) \strong{Label} (column 1): The type of each sample ('Sample' or 'Blank);
2) \strong{Sample} (column 2): The name of each sample;
3) \strong{Tag} (column 3 to ...): Tags for connecting samples and blanks. Samples and their
corresponding blanks should have the same tag in one column. More than one columns can be used.}

\item{div_step}{The step size for library division.}

\item{div_max_lib}{The maximum library number.}

\item{div_min_lib}{The minimum library number.}

\item{div_num_lib}{Apart from the maximum and minimum library number, user can also set a fixed library number.}

\item{div_print_summary}{When \code{div_print_summary} is \code{TRUE}, the function will print the statistic information
for each library design scheme.}
}
\value{
A list contains:
1) \strong{Blank_matrix}: Standard sample-blank matrix.
2) \strong{Blank_index}: A table contains all samples (column 1) and their corresponding blanks (column 2).
3) \strong{Library}: A list containing the detailed information of each library scheme named as 'Library_x'
(x: the library number). Under each library number, a library scheme can contain multiple specific
secondary schemes named as 'Case_y' (y: ID of secondary scheme). The information of each secondary
scheme including:
a) \strong{Start_index}: The start sample index of each library.
b) \strong{Cut_index}: The division index of each library, that is, \strong{Start_index - 1}.
c) \strong{lib_data}: The information of each library including its start (\code{Start}) and end (\code{End}) index,
numbers of samples (\code{nSample}), blanks (\code{nBlank}), their sums (\code{nTotal}) and ratios
(\code{SBR} = \code{nSample}/\code{nBlank}).
4) \strong{Library_summary}: The summary of all library schemes, including the total sample numbers (including blanks)
of each library and the maximum, minimum and mean value of the total sample number and SBR across these libraries.
}
\description{
Function \code{Lib_design} generate different library design schemes based on original sample list and
calculate basic information for each library including the number of samples and the ratio
of samples to blanks.
}
\examples{
data_test <- data(sample_list_test)
cat('Sample list example:')
head(data_test)

lib_dsg_res <- Lib_design(sample_list_test, div_step=8, div_max_lib=3, div_min_lib=1)
cat('Standard sample-blank matrix:')
head(lib_dsg_res[['Blank_matrix']])
cat('Sample-blank table:')
head(lib_dsg_res[['Blank_index']])
cat('Library division summary:')
print(lib_dsg_res[['Library_summary']])
}
