% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_master.R
\name{Lib_division}
\alias{Lib_division}
\title{Divide samples into libraries.}
\usage{
Lib_division(
  blank_matrix,
  step = 8,
  max_lib = 2,
  min_lib = 1,
  num_lib = NULL,
  print_summary = FALSE
)
}
\arguments{
\item{blank_matrix}{A standard sample-blank matrix.}

\item{step}{The step size for library division.}

\item{max_lib}{The maximum library number.}

\item{min_lib}{The minimum library number.}

\item{num_lib}{Apart from the maximum and minimum library number, user can also set a fixed library number.}

\item{print_summary}{When this parameter is 'TRUE', the function will print the final statistic information
for each library design scheme.}
}
\value{
A list containing:
1) \strong{Blank_matrix}: the sample-blank matrix as input.
2) \strong{Library}: A list containing the detailed information of each library scheme named as \code{Library_x}
(x: the library number). Under each library number, a library scheme can contain multiple specific
secondary schemes named as \code{Case_y} (y: ID of secondary scheme). The information of each secondary
scheme including:
a) \strong{Start_index}: The start sample index of each library.
b) \strong{Cut_index}: The division index of each library, that is, \strong{Start_index - 1}.
c) \strong{lib_data}: The information of each library including its start (\code{Start}) and end (\code{End}) index,
numbers of samples (\code{nSample}), blanks (\code{nBlank}), their sums (\code{nTotal}) and ratios
(\code{SBR} = \code{nSample}/\code{nBlank}).
3) \strong{Summary}: The summary of all library schemes, including the total sample numbers (including blanks)
of each library and the maximum, minimum and mean value of the total sample number and SBR across these libraries.
}
\description{
Function \code{Lib_division} divides samples into multiple libraries and export basic information of each library
under each division case.
}
\examples{
set.seed(2026)
blank_mt_test <- data.frame('Sample'=paste0('S', 1:32),
                            'B1'=sample(0:1, 32, replace = T, prob=c(0.75,0.25)),
                            'B2'=sample(0:1, 32, replace = T, prob=c(0.75,0.25)),
                            'B3'=sample(0:1, 32, replace = T, prob=c(0.75,0.25)),
                            'B4'=sample(0:1, 32, replace = T, prob=c(0.75,0.25)))
cat('Standard sample-blank matrix example:')
head(blank_mt_test)

lib_div_res <- Lib_division(blank_mt_test, step=8, max_lib=3, min_lib=1, print_summary=F)
cat('Standard sample-blank matrix:')
head(lib_div_res[['Blank_matrix']])
cat('Library division summary:')
print(lib_div_res[['Summary']])

}
