% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_master.R
\name{Library_design}
\alias{Library_design}
\title{Generate library design schemes}
\usage{
Library_design(
  sample_list,
  div_step = 8,
  div_max_lib = 2,
  div_min_lib = 1,
  div_num_lib = NULL,
  div_print_summary = FALSE
)
}
\arguments{
\item{sample_list}{The original sample and blank list as data.frame, with each sample as one row.
It should include:
\enumerate{
\item \strong{Label} (column 1): The type of each sample ('Sample' or 'Blank);
\item \strong{Sample} (column 2): The name of each sample;
\item \strong{Tag} (column 3 ~ ...): Tags for connecting samples and blanks. Samples and their corresponding blanks should have
the same tag in one column. More than one columns can be used.
}}

\item{div_step}{The step size for library division.}

\item{div_max_lib}{The maximum library number.}

\item{div_min_lib}{The minimum library number.}

\item{div_num_lib}{Apart from the maximum and minimum library number, user can also set a fixed library number.}

\item{div_print_summary}{When this parameter is set as 'TRUE', the function will print the final statistic information
for each library design scheme.}
}
\value{
A list contains:
\enumerate{
\item \strong{Blank_matrix}: Standard sample-blank matrix.
\item \strong{Blank_index}: A table contains all samples (column 1) and their corresponding blanks (column 2).
\item \strong{Library}: A list containing the detailed information of each library scheme named as 'Library_x' (x: the library number).
Under each library number, a library scheme can contain multiple specific secondary schemes named as 'Case_y' (y: ID of
secondary scheme). The information of each secondary scheme including:
a) \strong{Start_index}: The start sample index of each library.
b) \strong{Cut_index}: The division index of each library, that is, \strong{Start_index - 1}.
c) \strong{lib_data}: The information of each library including its start (Start) and end (End) index, numbers of samples (nSample),
blanks (nBlank), their sums (nTotal) and ratios (SBR = nSample/nBlank).
\item \strong{Library_summary}: The summary of all library schemes, including the total sample numbers (including blanks) of each library
and the maximum, minimum and mean value of the total sample number and SBR across these libraries.
}
}
\description{
This function generate different library design schemes based on original sample list and
calculate basic information for each library including the number of samples and the ratio of samples to
blanks.
}
\examples{
data_test <- data(sample_list_test)
lib_dsg_res <- Library_design(sample_list_test, div_step=8, div_max_lib=3, div_min_lib=1)
cat('Standard sample-blank matrix:')
head(lib_dsg_res[['Blank_matrix']])
cat('Sample-blank table:')
head(lib_dsg_re[['Blank_index']])
cat('Library division summary:')
head(lib_dsg_re[['Library_summary']])
}
