% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_master.R
\name{PCR_list}
\alias{PCR_list}
\title{Generate PCR list and 96-well PCR plates using sample-blank matrix}
\usage{
PCR_list(
  blank_mt,
  cut_index = 1,
  tail = NULL,
  include_PB = TRUE,
  primer_list = NULL,
  primer_index = NULL,
  byID = "PCR_ID",
  print_list = FALSE,
  print_plate = FALSE,
  print_pcrnum = FALSE
)
}
\arguments{
\item{blank_mt}{Standard sample-blank matrix.}

\item{cut_index}{A vector containing the start index of each library in the sample-blank matrix.}

\item{tail}{If \code{tail} is not \code{NULL}, this function will add it after each \code{PCR_ID} in the PCR list.}

\item{include_PB}{If \code{include_PB} is \code{TRUE}, this function will add 3 PCR blanks (PB-1/2/3) in the PCR list.}

\item{primer_list}{The primer list (data.frame) for assigning primers to each PCR. It is suggested to contain the sequence
and ID of each primer.}

\item{primer_index}{The start primer index for assignment in the primer list. If \code{primer_index} is \code{NULL},
the primer assignment will start with the first primer.}

\item{byID}{Name of column (character) which is used to be shown in the final PCR plates. This columnn
should distinguish different PCR replications.}

\item{print_list}{If \code{print_list} is \code{TRUE}, this function will export PCR list as an excel file (.xlsx) under the
working directory.}

\item{print_plate}{If \code{print_plate} is \code{TRUE}, this function will export PCR plates as an excel file (.xlsx) under the
working directory.}

\item{print_pcrnum}{If \code{print_pcrnum} is \code{TRUE}, this function will export PCR number information as an excel file
(.xlsx) under the working directory.}
}
\value{
This function returns a list containing:
1) \strong{Library-X}: The sample-blank matrix, PCR list and PCR plates for library X (X: ID number for each library);
2) \strong{PCR_number}: The number information for each library, including the presence/absence of each sample
and the sum sample number for each library.
}
\description{
Generate PCR list and 96-well PCR plates using sample-blank matrix
}
\examples{
set.seed(2026)
blank_mt_test <- data.frame('Sample'=paste0('S', 1:32),
                            'B1'=sample(0:1, 32, replace = T, prob=c(0.75,0.25)),
                            'B2'=sample(0:1, 32, replace = T, prob=c(0.75,0.25)),
                            'B3'=sample(0:1, 32, replace = T, prob=c(0.75,0.25)),
                            'B4'=sample(0:1, 32, replace = T, prob=c(0.75,0.25)))
cat("Standard sample-blank matrix:")
print(blank_mt_test)

primer_test <- data.frame('PrimerID'=paste0('Primer-',1:100), 'Sequence'=NA)
for (i in 1:100){
  primer_test$Sequence[i] <- paste0(sample(c('a','c','t','g'), 6, replace = T), collapse = '')
}
cat("Primer list:")
head(primer_test)

pcr_list_res <- PCR_list(blank_mt_test, cut_index=c(1,17), include_PB = TRUE,
                         primer_list = primer_test, primer_index = c(1,20), byID='PCR_ID',
                         print_list = FALSE,  print_plate = FALSE, print_pcrnum = FALSE)
cat("PCR list:")
head(pcr_list_res[[1]][['PCR_list']])
cat("PCR plate:")
head(pcr_list_res[[1]][['Plate']])
cat("PCR number information:")
head(pcr_list_res[['PCR_number']])

}
