% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_master.R
\name{PCR_table}
\alias{PCR_table}
\title{Generate 96-well PCR plates based on PCR list.}
\usage{
PCR_table(pcr_list, by.id = NULL)
}
\arguments{
\item{pcr_list}{A data frame containing PCR information. It should \strong{at least} contains:
1) \strong{Sample}: Sample ID used to distinguish different samples.
2) \strong{PCR_ID}: PCR ID used to distinguish different PCR replications for the same sample
(such as "X-1", "X-2", "X-3" for the sample "X"). The name of this column must correspond
with the parameter \code{by.id}.}

\item{by.id}{Character, the name of column which is used to be shown in the final PCR plates. This columnn
should distinguish different PCR replications.}
}
\value{
A list containing the information of all 96-well PCR plates named as "PLATE-x" (x: number of each plate).
Each plate is a 12-row × 8-column data frame corresponding to the layout of a 96-well plate, filled with
PCR ID by \strong{by.id} column. Every 3 rows corresponds to 8 samples (i.e., each sample occupies 3 rows × 1 column).
Specially, if \code{by.id} is \code{NULL}, returns \code{0}.
}
\description{
Generate 96-well PCR plates based on PCR list.
}
\examples{
data_test <- data.frame('Sample'=paste0('S', rep(1:32, each=3)),
                        'PCR_ID'=paste0('S', rep(1:32, each=3), '-', rep(1:3, each=32)))
cat("PCR list example:")
head(data_test)

pcr_table_res <- PCR_table(data_test, by.id='PCR_ID')
cat("Plate template:")
print(pcr_table_res[[1]])

}
