% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_list_test}
\alias{sample_list_test}
\title{Sample list data}
\format{
A tibble data frame with 71 rows and 4 columns：
\describe{
\item{Lable}{character, type of each sample ('Sample' or 'Blank')}
\item{ID}{character, unique sample identifiers}
\item{Date_1}{character, the 1st processing date in YYYYMMDD format}
\item{Date_2}{character, the 2nd processing date in YYYYMMDD format}
}
}
\source{
Laboratory sample processing records
}
\usage{
sample_list_test
}
\description{
An example sample data containing 54 samples and 17 blanks.
}
\keyword{datasets}
