% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_master.R
\name{Blank_statistic}
\alias{Blank_statistic}
\title{Record and calculate the basic information of a library based on standard sample-blank matrix}
\usage{
Blank_statistic(blank_mt, start = NA, end = NA)
}
\arguments{
\item{blank_mt}{A standard sample-blank matrix.}

\item{start}{The row index of the start sample in \strong{blank_mt}.}

\item{end}{The row index of the ending sample in \strong{blank_mt}.}
}
\value{
A vector contains:
\enumerate{
\item Number of samples (\strong{nSample});
\item Number of blanks (\strong{nBlank});
\item Sum number of samples and blanks (\strong{nTotal});
\item Ratio of \strong{nSample} to \strong{nBlank}.
}
}
\description{
Function \strong{Blank_statistic} records the start and ending sample of a library in a standard sample-blank matrix,
and calculates the number of samples and blanks and the sample-blank ratio (SBR, sample/blank).
}
\examples{
blank_mt_test <- data.frame('Sample'=c('S1', 'S2', 'S3', 'S4'),
'B1'=c(4,1,0,0),
'B2'=c(0,0,1,1),
'B3'=c(1,1,0,0),
'B4'=c(0,0,0,1))
blank_sta_res <- Blank_statistic(blank_mt_test, start=2, end=2)
print(blank_sta_res)
}
